Move default context to a standalone file
